Ho implementato delle modifiche all'smtracer per tener conto degli accessi effettuati nei blocchi non selezionati dalla discplina di instrumentazione quando si è in modalità simulata. Questo dovrebbe completare le modifiche al tracer, a meno dell'insorgere di qualche bug e di dover modificare l'euristica di selezione degli accessi.

+++ src/presets/smtracer/smtracer.h
@@ -47,8 +47,8 @@ typedef struct smt_access {
 
   struct smt_access *original;    // Pointer to the equal original access
 
-  char vtable[SMT_VTABLE_SIZE];       // Version table for general-purpose registers
-  float score;                    // Access instrumentation score
+  char vtable[SMT_VTABLE_SIZE];   // Version table for general-purpose registers
+  double score;                   // Access instrumentation score
 
   struct smt_access *next;
 } smt_access;
@@ -58,9 +58,9 @@ typedef struct {
   // Block-level features
   block *lheader;             // Closest loop header
   unsigned int cycles;        // Number of joined program cycles
-  float memratio;             // Memory sensitivity
+  double memratio;            // Memory sensitivity
 
-  float score;                // Relative score ranging in [0,1]
+  double score;               // Relative score ranging in [0,1]
 
   smt_access *candidates;     // Candidates list
